scee/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
argon2-cffi==23.1.0
Werkzeug==3.0.1
